feat: 게시글 작성, 조회 시 DB 접근 기술 변경

+++ src/main/java/kim/jiwook/playground/controller/restController/RestBoardController.java
@@ -29,7 +29,7 @@ public class RestBoardController {
      * @throws SQLException
      */
     @PostMapping("/board")
-    public ResponseEntity<Integer> restfulInsertBoard(@RequestBody RequestInsertBoard requestVO) throws IOException, SQLException {
+    public ResponseEntity<Long> restfulInsertBoard(@RequestBody RequestInsertBoard requestVO) throws IOException, SQLException {
         return ResponseEntity.ok(boardService.insertBoard(requestVO));
     }
